Update docs, readme; add swift package index file; remove jazzy file
