--- .jazzy.yaml
@@ -1,20 +1,0 @@
-module: Relax
-github_url: https://github.com/tdeleon/Relax
-copyright: © 2022 Thomas De Leon, all rights reserved.
-clean: true
-output: docs
-hide_documentation_coverage: true
-documentation: Documentation/*.md
-custom_categories:
-  - name: Defining Services
-    children:
-    - Service
-    - Endpoint
-  - name: Making Requests
-    children:
-    - ServiceRequest
-    - HTTPRequestMethod
-    - RequestContentType
-    - RequestError
-abstract: Documentation/Sections/*.md
-
